+++ preprocessor/pre-post/ecs_pre_gmsh.c
@@ -51,6 +51,9 @@
  *----------------------------------------------------------------------------*/
 
 #include "ecs_descr.h"
+#include "ecs_descr_chaine.h"
+#include "ecs_famille.h"
+#include "ecs_famille_chaine.h"
 #include "ecs_maillage.h"
 #include "ecs_maillage_priv.h"
 
@@ -83,8 +86,9 @@
 
 /* Valeurs associées aux dimensionnements pour la lecture des lignes */
 
-#define ECS_LOC_LNG_MAX_CHAINE_GMSH 514 /* Dimension des chaînes de réception
-                                           des lignes lues */
+#define ECS_LOC_LNG_MAX_CHAINE_GMSH  2048 /* Max file line length */
+
+typedef int gmsh_int_t;
 
 /*============================================================================
  *                          Définitions de types
@@ -368,7 +372,7 @@ ecs_loc_pre_gmsh__lit_vers_format(ecs_file_t     *fic_maillage,
     int   un;
 
     ecs_file_set_type(fic_maillage, ECS_FILE_TYPE_BINARY);
-    ecs_file_read(&un, 4, 1, fic_maillage);
+    ecs_file_read(&un, sizeof(gmsh_int_t), 1, fic_maillage);
     if (un != 1)
       ecs_file_set_swap_endian(fic_maillage, 1);
 
@@ -472,11 +476,13 @@ ecs_loc_pre_gmsh__lit_nodes(ecs_maillage_t   *maillage,
 
     for (ind_nod = 0; ind_nod < nbr_nod; ind_nod++) {
 
+      gmsh_int_t _label;
       double xyz[3];
 
-      ecs_file_read(&((*som_val_label)[ind_nod]), sizeof(int), 1, fic_maillage);
+      ecs_file_read(&_label, sizeof(gmsh_int_t), 1, fic_maillage);
       ecs_file_read(&xyz, sizeof(double), 3, fic_maillage);
 
+      (*som_val_label)[ind_nod] = (ecs_int_t)_label;
       for (icoo = 0; icoo < 3; icoo++)
         som_val_coord[ind_nod * 3 + icoo] = xyz[icoo];
 
@@ -508,6 +514,353 @@ ecs_loc_pre_gmsh__lit_nodes(ecs_maillage_t   *maillage,
 }
 
 /*----------------------------------------------------------------------------
+ * Read mesh physical names
+ *----------------------------------------------------------------------------*/
+
+static void
+_read_physical_names(ecs_file_t       *mesh_file,
+                     int              *line_num,
+                     int              *n_phys_names,
+                     int              *n_phys_names_max,
+                     int             **phys_name_id,
+                     char           ***phys_name)
+{
+  char   line[ECS_LOC_LNG_MAX_CHAINE_GMSH];
+  char   name[ECS_LOC_LNG_MAX_CHAINE_GMSH];
+  int    retval;
+
+  int    cpt_phys_names = 0;
+  int    _n_phys_names_max = -1;
+
+  int    _n_phys_names = 0;
+  int   *_phys_name_id = NULL;
+  char **_phys_name = NULL;
+
+  /*xxxxxxxxxxxxxxxxxxxxxxxxxxx Instructions xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx*/
+
+  ecs_file_gets(line, ECS_LOC_LNG_MAX_CHAINE_GMSH, mesh_file, line_num);
+
+  retval = sscanf(line,"%d", &_n_phys_names);
+
+  if (retval != 1)
+    ecs_error(__FILE__, __LINE__, 0,
+              _("Error decoding line %ld of file\n\"%s\" :\n"
+                "The information on physical groups form "
+                "\"numPhysicalNames\"."),
+                (long)(*line_num), ecs_file_get_name(mesh_file));
+
+  for (int i = 0; i < _n_phys_names; i++) {
+
+    gmsh_int_t  dim_phys_name, tag_phys_name;
+
+    ecs_file_gets(line,
+                  ECS_LOC_LNG_MAX_CHAINE_GMSH,
+                  mesh_file,
+                  line_num);
+
+    retval = sscanf(line, "%d %d %s",
+                    &dim_phys_name, &tag_phys_name, name);
+
+    if (retval != 3)
+      ecs_error
+        (__FILE__, __LINE__, 0,
+         _("Error decoding line %ld of file\n\"%s\" :\n"
+           "The description of entity <%d> physical group was expected in the form "
+           "\"dimension tag names\"."),
+           (long)(*line_num), ecs_file_get_name(mesh_file));
+
+    if (dim_phys_name > 1) {
+      if (tag_phys_name > _n_phys_names_max) {
+        ECS_REALLOC(_phys_name_id, tag_phys_name + 1, int);
+        for (int j = _n_phys_names_max+1; j < tag_phys_name + 1; j++)
+          _phys_name_id[j] = -1;
+        _n_phys_names_max = tag_phys_name;
+      }
+      _phys_name_id[tag_phys_name] = cpt_phys_names;
+
+      ECS_REALLOC(_phys_name, cpt_phys_names+1, char *);
+      ECS_MALLOC(_phys_name[cpt_phys_names], strlen(name) + 1, char);
+      strcpy(_phys_name[cpt_phys_names], name);
+
+      cpt_phys_names += 1;
+    }
+
+  }
+
+  /* End of section line */
+
+  ecs_file_gets(line, ECS_LOC_LNG_MAX_CHAINE_GMSH, mesh_file, line_num);
+
+  if (strncmp(line, "$EndPhysicalNames", strlen("$EndPhysicalNames")) != 0)
+    ecs_error(__FILE__, __LINE__, 0,
+              _("The end of the physical names specification of file\n"
+                "\"%s\" is not present at the expected place.)"),
+              ecs_file_get_name(mesh_file));
+
+  /* Return values */
+
+  *n_phys_names = cpt_phys_names;
+  *n_phys_names_max = _n_phys_names_max;
+  *phys_name_id = _phys_name_id;
+  *phys_name = _phys_name;
+}
+
+/*----------------------------------------------------------------------------
+ * Read entities
+ *----------------------------------------------------------------------------*/
+
+static void
+_read_entities(int             n_phys_names_max,
+               const int       phys_name_id[],
+               const char    **phys_name,
+               ecs_file_t     *mesh_file,
+               int            *line_num,
+               int             type_fmt_gmsh,
+               int             fam_shift[ECS_N_ENTMAIL],
+               ecs_famille_t  *vect_famille_tete[ECS_N_ENTMAIL])
+{
+  char   line[ECS_LOC_LNG_MAX_CHAINE_GMSH];
+  int    retval;
+
+  unsigned long  _n_elts[4] = {0, 0, 0, 0};
+
+  /*xxxxxxxxxxxxxxxxxxxxxxxxxxx Instructions xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx*/
+
+  /* Initialize families */
+
+  for (int l = 0; l < ECS_N_ENTMAIL; l++) {
+    vect_famille_tete[l] = NULL;
+    fam_shift[l] = 0;
+  }
+
+  /* Read based on format */
+
+  if (type_fmt_gmsh == 0) {
+
+    ecs_file_gets(line, ECS_LOC_LNG_MAX_CHAINE_GMSH, mesh_file, line_num);
+
+    retval = sscanf(line,"%lu %lu %lu %lu",
+                    &(_n_elts[0]), &(_n_elts[1]), &(_n_elts[2]), &(_n_elts[3]));
+
+    if (retval != 4)
+      ecs_error(__FILE__, __LINE__, 0,
+                _("Error decoding line %ld of file\n\"%s\" :\n"
+                  "The information on entitiess form "
+                  "\"numPoints numCurves numSurfaces numVolumes\"."),
+                (long)(*line_num), ecs_file_get_name(mesh_file));
+
+  }
+  else {
+
+    ecs_file_set_type(mesh_file, ECS_FILE_TYPE_BINARY);
+
+    ecs_file_read(_n_elts, sizeof(unsigned long), 4, mesh_file);
+
+  }
+
+  /* Loop on entity type */
+
+  int ifam_base = 0;
+
+  for (int l = 0; l < 4; l++) {
+
+    int  ifam_max = 0;
+    ecs_int_t  ifam_ent = -1;
+
+    if (l == 2)
+      ifam_ent = ECS_ENTMAIL_FAC;
+    else if (l == 3)
+      ifam_ent = ECS_ENTMAIL_CEL;
+
+    if (ifam_ent > -1)
+      fam_shift[ifam_ent] = ifam_base;
+
+    unsigned int  n = _n_elts[l];
+
+    for (unsigned long k = 0; k < n; k++) {
+
+      int err = 0;
+
+      /* Text variant */
+
+      if (type_fmt_gmsh == 0) {
+
+        ecs_file_gets(line, ECS_LOC_LNG_MAX_CHAINE_GMSH, mesh_file, line_num);
+
+        char *ssch = strtok(line, " ");
+        if (ssch == NULL) {
+          err = 1;
+          break;
+        }
+
+        int tag = atoi(ssch);
+
+        if (ifam_ent > -1 && tag > ifam_max)
+          ifam_max = tag;
+
+        for (int m = 0; m < 6; m++) { /* bounding box (unused here) */
+          ssch = strtok(NULL, " ");
+          if (ssch == NULL) {
+            err = 1;
+            break;
+          }
+        }
+
+        ssch = strtok(NULL, " ");
+        if (ssch == NULL) {
+          err = 1;
+          break;
+        }
+
+        unsigned long n_phys = atol(ssch);
+
+        ecs_descr_t  *descr_tete = NULL;
+
+        for (unsigned long m = 0; m < n_phys; m++) {
+          ssch = strtok(NULL, " ");
+          if (ssch == NULL) {
+            err = 1;
+            break;
+          }
+          int phys_tag = atoi(ssch);
+          if (ifam_ent > -1) {
+            ecs_descr_t *descr = ecs_descr__cree(phys_tag, NULL);
+            ecs_descr_chaine__ajoute(&descr_tete, descr);
+            if (phys_tag > -1 && phys_tag <= n_phys_names_max) {
+              int id = phys_name_id[phys_tag];
+              if (id > -1) {
+                const char *name = phys_name[id];
+                descr = ecs_descr__cree(ECS_DESCR_IDE_NUL, name);
+                ecs_descr_chaine__ajoute(&descr_tete, descr);
+              }
+            }
+          }
+        }
+
+        if (ifam_ent > -1) {
+          ecs_famille_t *fml = ecs_famille__cree(tag + ifam_base, descr_tete);
+          ecs_famille_chaine__ajoute(&vect_famille_tete[ifam_ent], fml);
+        }
+
+        if (l > 0) { /* bounding entities */
+
+          ssch = strtok(NULL, " ");
+          if (ssch == NULL) {
+            err = 1;
+            break;
+          }
+
+          unsigned long n_bound = atol(ssch);
+          for (unsigned long m = 0; m < n_bound; m++) {
+            ssch = strtok(NULL, " ");
+            if (ssch == NULL) {
+              err = 1;
+              break;
+            }
+          }
+
+        }
+
+        if (err > 0)
+          ecs_error(__FILE__, __LINE__, 0,
+                    _("Error decoding line %ld of file\n\"%s\" :\n"
+                      "The information on entities does not seem to match the spec."),
+                    (long)(*line_num), ecs_file_get_name(mesh_file));
+
+      }
+      else { /* Binary variant */
+
+        int tag = 0;
+        ecs_file_read(&tag, sizeof(int), 1, mesh_file);
+
+        if (ifam_ent > -1 && tag > ifam_max)
+          ifam_max = tag;
+
+        /* bounding box (unused here) */
+        double bbox[6];
+        ecs_file_read(bbox, sizeof(double), 6, mesh_file);
+
+        unsigned long n_phys = 0;
+        ecs_file_read(&n_phys, sizeof(unsigned long), 1, mesh_file);
+
+        ecs_descr_t  *descr_tete = NULL;
+
+        for (unsigned long m = 0; m < n_phys; m++) {
+          int phys_tag = 0;
+          ecs_file_read(&phys_tag, sizeof(int), 1, mesh_file);
+          if (ifam_ent > -1) {
+            ecs_descr_t *descr = ecs_descr__cree(phys_tag, NULL);
+            ecs_descr_chaine__ajoute(&descr_tete, descr);
+            if (phys_tag > -1 && phys_tag <= n_phys_names_max) {
+              int id = phys_name_id[phys_tag];
+              if (id > -1) {
+                const char *name = phys_name[id];
+                descr = ecs_descr__cree(ECS_DESCR_IDE_NUL, name);
+                ecs_descr_chaine__ajoute(&descr_tete, descr);
+              }
+            }
+          }
+        }
+
+        if (ifam_ent > -1) {
+          ecs_famille_t *fml = ecs_famille__cree(tag + ifam_base, descr_tete);
+          ecs_famille_chaine__ajoute(&vect_famille_tete[ifam_ent], fml);
+        }
+
+        if (l > 0) { /* bounding entities */
+
+          unsigned long n_bound = 0;
+          ecs_file_read(&n_bound, sizeof(unsigned long), 1, mesh_file);
+
+          for (unsigned long m = 0; m < n_bound; m++) {
+            int bound_tag;
+            ecs_file_read(&bound_tag, sizeof(int), 1, mesh_file);
+          }
+
+        }
+
+      } /* End of test on format */
+
+    } /* End of loop on elements */
+
+    if (ifam_ent > -1) {
+      int j = 0;
+      ifam_max += ifam_base;
+      while (ifam_max > 0) {
+        j++;
+        ifam_max /= 10;
+      }
+      ifam_max = 1;
+      while (j > 0) {
+        ifam_max *= 10;
+        j--;
+      }
+      ifam_base = ifam_max;
+    }
+
+  } /* end of loop on entity types */
+
+  /* End of section line */
+
+  if (type_fmt_gmsh != 0) {
+    ecs_file_set_type(mesh_file, ECS_FILE_TYPE_TEXT);
+    ecs_file_gets(line, 2, mesh_file, line_num);
+  }
+
+  ecs_file_gets(line, ECS_LOC_LNG_MAX_CHAINE_GMSH, mesh_file, line_num);
+
+  if (strncmp(line, "$EndEntities", strlen("$EndEntities")) != 0)
+    ecs_error(__FILE__, __LINE__, 0,
+              _("The end of the entities specification of file\n"
+                "\"%s\" is not present at the expected place.)"),
+              ecs_file_get_name(mesh_file));
+
+  /* Return values */
+
+  return vect_famille_tete;
+}
+
+/*----------------------------------------------------------------------------
  *  Lecture des coordonnées des noeuds
  *----------------------------------------------------------------------------*/
 
@@ -523,7 +876,6 @@ ecs_loc_pre_gmsh__lit_nodes_v4(ecs_maillage_t   *maillage,
 
   /* Variables Gmsh lues */
 
-  long           label;
   double         coord[3];
   unsigned long  n_ent_blocks, nbr_nod;
 
@@ -588,7 +940,8 @@ ecs_loc_pre_gmsh__lit_nodes_v4(ecs_maillage_t   *maillage,
                       fic_maillage,
                       num_ligne);
 
-        retour = sscanf(chaine,"%ld %lg %lg %lg",
+        int label;
+        retour = sscanf(chaine,"%d %lg %lg %lg",
                         &label, coord, coord+1, coord+2);
 
         if (retour != 4)
@@ -630,7 +983,7 @@ ecs_loc_pre_gmsh__lit_nodes_v4(ecs_maillage_t   *maillage,
 
     for (unsigned long eb = 0; eb < n_ent_blocks; eb++) {
 
-      unsigned long buf_i[3];
+      int buf_i[3];
       unsigned long n_ent_nodes;
 
       ecs_file_read(buf_i, sizeof(int), 3, fic_maillage);
@@ -638,11 +991,13 @@ ecs_loc_pre_gmsh__lit_nodes_v4(ecs_maillage_t   *maillage,
 
       for (unsigned long bn = 0; bn < n_ent_nodes; bn++, ind_nod++) {
 
+        int label;
         double xyz[3];
 
-        ecs_file_read(&((*som_val_label)[ind_nod]), sizeof(int), 1, fic_maillage);
+        ecs_file_read(&label, sizeof(int), 1, fic_maillage);
         ecs_file_read(&xyz, sizeof(double), 3, fic_maillage);
 
+        (*som_val_label)[ind_nod] = label;
         for (icoo = 0; icoo < 3; icoo++)
           som_val_coord[ind_nod * 3 + icoo] = xyz[icoo];
 
@@ -706,8 +1061,8 @@ ecs_loc_pre_gmsh__lit_elements(ecs_maillage_t  *maillage,
   ecs_int_t   ind_tag_elt;
   ecs_int_t   nbr_tag_elt;
 
-  ecs_int_t   header[3];
-  ecs_int_t   data[100];
+  gmsh_int_t  header[3];
+  gmsh_int_t  data[100];
 
   int         type_gmsh;
   ecs_int_t   nbr_nod_elt_gmsh;
@@ -994,7 +1349,7 @@ ecs_loc_pre_gmsh__lit_elements(ecs_maillage_t  *maillage,
 
     while (nbr_elt_lus != nbr_elt_tot) {
 
-      ecs_file_read(&header, 4, 3, fic_maillage);
+      ecs_file_read(&header, sizeof(gmsh_int_t), 3, fic_maillage);
 
       type_gmsh   = header[0];
       nbr_elt     = header[1];
@@ -1012,7 +1367,7 @@ ecs_loc_pre_gmsh__lit_elements(ecs_maillage_t  *maillage,
 
       for (ind_elt = 0; ind_elt < nbr_elt; ind_elt++) {
 
-        ecs_file_read(&data, sizeof(int),
+        ecs_file_read(&data, sizeof(gmsh_int_t),
                       1 + nbr_tag_elt + nbr_nod_elt_gmsh, fic_maillage);
 
         if (type_gmsh == GMSH_POINT1) {
@@ -1103,15 +1458,15 @@ ecs_loc_pre_gmsh__lit_elements(ecs_maillage_t  *maillage,
   /* Indications */
 
   if (cpt_point > -1 || cpt_are > 0)
-      printf("\n");
+    printf("\n");
   if (cpt_point > -1)
-      printf(_("  %ld elements of type \"point\" ignored\n"),
-             (long)cpt_point);
+    printf(_("  %ld elements of type \"point\" ignored\n"),
+           (long)cpt_point);
   if (cpt_are > -1)
-      printf(_("  %ld elements of type \"edge\" ignored\n"),
-             (long)cpt_are);
+    printf(_("  %ld elements of type \"edge\" ignored\n"),
+           (long)cpt_are);
   if (cpt_point > -1 || cpt_are > 0)
-      printf("\n");
+    printf("\n");
 
   /* Ligne de fin de rubrique */
 
@@ -1151,7 +1506,6 @@ ecs_loc_pre_gmsh__lit_elements(ecs_maillage_t  *maillage,
 
   ECS_FREE(*som_val_label);
 
-
   /* Transfert des valeurs lues dans les structures d'entités de maillage */
   /*======================================================================*/
 
@@ -1176,6 +1530,7 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
                                   int              type_fmt_gmsh,
                                   int             *num_ligne,
                                   ecs_int_t        nbr_som,
+                                  int              fam_shift[ECS_N_ENTMAIL],
                                   ecs_int_t      **som_val_label)
 {
   char        chaine[ECS_LOC_LNG_MAX_CHAINE_GMSH];
@@ -1188,26 +1543,20 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
 
   long        label;
   ecs_int_t   type_ecs;
-  ecs_int_t   coul_elt;
   ecs_int_t   ind_elt;
   ecs_int_t   nbr_elt_lus;
   ecs_int_t   ind_nod_elt;
   ecs_int_t   ind_som_elt;
   ecs_int_t   nbr_som_elt;
 
-  ecs_int_t   data[100];
+  int         data[100];
 
   int         type_gmsh;
   ecs_int_t   nbr_nod_elt_gmsh;
   ecs_int_t   num_nod_elt_gmsh[ECS_GMSH_NBR_MAX_SOM];
   ecs_int_t   ent_num;
 
-  ecs_int_t   icoul;
   ecs_int_t   ient;
-
-  ecs_int_t    cpt_coul_ent[ECS_N_ENTMAIL]; /* Compteur de couleurs         */
-  ecs_int_t   *val_coul_ent[ECS_N_ENTMAIL]; /* Tableau valeurs des couleurs */
-  ecs_size_t  *cpt_elt_coul_ent[ECS_N_ENTMAIL];
 
   /* Stockage avant transfert */
   /*--------------------------*/
@@ -1216,7 +1565,7 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
 
   ecs_size_t  *elt_pos_som_ent    [ECS_N_ENTMAIL]; /* Positions numeros som */
   ecs_int_t   *elt_val_som_ent    [ECS_N_ENTMAIL]; /* Numeros des sommets   */
-  ecs_int_t   *elt_val_color_ent  [ECS_N_ENTMAIL]; /* Couleurs              */
+  int         *elt_val_fam_ent    [ECS_N_ENTMAIL]; /* Groupes (familles)    */
 
   ecs_int_t cpt_are   = 0;
   ecs_int_t cpt_point = 0;
@@ -1234,13 +1583,7 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
   /* `ECS_ENTMAIL_DEB = ECS_ENTMAIL_SOM'      */
 
   for (ient = ECS_ENTMAIL_FAC; ient < ECS_N_ENTMAIL; ient++) {
-
-    cpt_elt_ent        [ient] = 0;
-
-    cpt_coul_ent       [ient] = 0   ;
-    cpt_elt_coul_ent   [ient] = NULL;
-    val_coul_ent       [ient] = NULL;
-
+    cpt_elt_ent[ient] = 0;
   }
 
   /* Décodage du nombre d'éléments (toutes dimensions confondues */
@@ -1278,8 +1621,8 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
 
   for (ient = ECS_ENTMAIL_FAC; ient < ECS_N_ENTMAIL; ient++) {
 
-    ECS_MALLOC(elt_val_color_ent[ient],   nbr_elt, ecs_int_t);
-    ECS_MALLOC(elt_pos_som_ent[ient],     nbr_elt + 1, ecs_size_t);
+    ECS_MALLOC(elt_val_fam_ent[ient], nbr_elt, int);
+    ECS_MALLOC(elt_pos_som_ent[ient], nbr_elt + 1, ecs_size_t);
 
     elt_pos_som_ent[ient][0] = 1;
   }
@@ -1322,7 +1665,6 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
            eb);
 
       type_gmsh = elt_type;
-      coul_elt = tag_ent;
 
       if (type_gmsh < (int)GMSH_SEG2 || type_gmsh > (int) GMSH_POINT1)
         ecs_error(__FILE__, __LINE__, 0,
@@ -1410,27 +1752,8 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
 
         /* Couleur (tag) de l'élément lu */
 
-        icoul = 0;
-        while (icoul < cpt_coul_ent[ent_num]           &&
-               val_coul_ent[ent_num][icoul] != coul_elt)
-          icoul++;
-
-        if (icoul == cpt_coul_ent[ent_num]) {
-
-          /* La valeur de la couleur n'a pas encore été stockée */
-
-          ECS_REALLOC(val_coul_ent[ent_num]    , cpt_coul_ent[ent_num] + 1,
-                      ecs_int_t);
-          ECS_REALLOC(cpt_elt_coul_ent[ent_num], cpt_coul_ent[ent_num] + 1,
-                      ecs_size_t);
-          cpt_elt_coul_ent[ent_num][icoul] = 0;
-          val_coul_ent[ent_num][icoul] = coul_elt;
-          cpt_coul_ent[ent_num]++;
-
-        }
-
-        cpt_elt_coul_ent[ent_num][icoul]++;
-        elt_val_color_ent[ent_num][cpt_elt_ent[ent_num]] = icoul + 1;
+        elt_val_fam_ent[ent_num][cpt_elt_ent[ent_num]]
+          = tag_ent + fam_shift[ent_num];
 
         /* Incrémentation du nombre d'éléments lus */
 
@@ -1464,7 +1787,7 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
       ecs_file_read(buf_i, sizeof(int), 3, fic_maillage);
       ecs_file_read(&n_ent_elt, sizeof(unsigned long), 1, fic_maillage);
 
-      coul_elt   = buf_i[0];
+      int tag_ent = buf_i[0];
       /* dim_ent = buf_i[1] */
       type_gmsh  = buf_i[2];
 
@@ -1492,7 +1815,7 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
           continue;
         }
 
-        label    = data[0];
+        label = data[0];
 
         /* Lecture des numéros des sommets de l'élément courant */
 
@@ -1508,8 +1831,8 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
 
         /* Position des numéros de sommets du prochain élément */
 
-        elt_pos_som_ent[ent_num][cpt_elt_ent[ent_num] + 1] =
-          elt_pos_som_ent[ent_num][cpt_elt_ent[ent_num]] + nbr_som_elt;
+        elt_pos_som_ent[ent_num][cpt_elt_ent[ent_num] + 1]
+          = elt_pos_som_ent[ent_num][cpt_elt_ent[ent_num]] + nbr_som_elt;
 
         /* Connectivité de l'élément par ses numéros de sommets */
 
@@ -1523,29 +1846,8 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
 
         }
 
-        /* Couleur (tag) de l'élément lu */
-
-        icoul = 0;
-        while (icoul < cpt_coul_ent[ent_num]           &&
-               val_coul_ent[ent_num][icoul] != coul_elt)
-          icoul++;
-
-        if (icoul == cpt_coul_ent[ent_num]) {
-
-          /* La valeur de la couleur n'a pas encore été stockée */
-
-          ECS_REALLOC(val_coul_ent[ent_num]    , cpt_coul_ent[ent_num] + 1,
-                      ecs_int_t);
-          ECS_REALLOC(cpt_elt_coul_ent[ent_num], cpt_coul_ent[ent_num] + 1,
-                      ecs_size_t);
-          cpt_elt_coul_ent[ent_num][icoul] = 0;
-          val_coul_ent[ent_num][icoul] = coul_elt;
-          cpt_coul_ent[ent_num]++;
-
-        }
-
-        cpt_elt_coul_ent[ent_num][icoul]++;
-        elt_val_color_ent[ent_num][cpt_elt_ent[ent_num]] = icoul + 1;
+        elt_val_fam_ent[ent_num][cpt_elt_ent[ent_num]]
+          = tag_ent + fam_shift[ent_num];
 
         /* Incrémentation du nombre d'éléments lus */
 
@@ -1565,15 +1867,15 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
   /* Indications */
 
   if (cpt_point > -1 || cpt_are > 0)
-      printf("\n");
+    printf("\n");
   if (cpt_point > -1)
-      printf(_("  %ld elements of type \"point\" ignored\n"),
-             (long)cpt_point);
+    printf(_("  %ld elements of type \"point\" ignored\n"),
+           (long)cpt_point);
   if (cpt_are > -1)
-      printf(_("  %ld elements of type \"edge\" ignored\n"),
-             (long)cpt_are);
+    printf(_("  %ld elements of type \"edge\" ignored\n"),
+           (long)cpt_are);
   if (cpt_point > -1 || cpt_are > 0)
-      printf("\n");
+    printf("\n");
 
   /* Ligne de fin de rubrique */
 
@@ -1595,12 +1897,10 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
   for (ient = ECS_ENTMAIL_FAC; ient < ECS_N_ENTMAIL; ient++) {
 
     if (cpt_elt_ent[ient] != 0) {
-      ECS_REALLOC(elt_pos_som_ent[ient]    ,
-                  cpt_elt_ent[ient] + 1    , ecs_size_t);
-      ECS_REALLOC(elt_val_som_ent[ient]    ,
+      ECS_REALLOC(elt_pos_som_ent[ient], cpt_elt_ent[ient] + 1, ecs_size_t);
+      ECS_REALLOC(elt_val_som_ent[ient],
                   elt_pos_som_ent[ient][cpt_elt_ent[ient]] - 1, ecs_int_t);
-      ECS_REALLOC(elt_val_color_ent[ient]  ,
-                  cpt_elt_ent[ient]        , ecs_int_t);
+      ECS_REALLOC(elt_val_fam_ent[ient], cpt_elt_ent[ient], int);
     }
 
     ecs_maillage_pre__label_en_indice
@@ -1610,9 +1910,7 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
        elt_val_som_ent[ient]);
   }
 
-
   ECS_FREE(*som_val_label);
-
 
   /* Transfert des valeurs lues dans les structures d'entités de maillage */
   /*======================================================================*/
@@ -1621,11 +1919,11 @@ ecs_loc_pre_gmsh__lit_elements_v4(ecs_maillage_t  *maillage,
                              cpt_elt_ent,
                              elt_pos_som_ent,
                              elt_val_som_ent,
+                             elt_val_fam_ent,
                              NULL,
-                             elt_val_color_ent,
-                             cpt_coul_ent,
-                             val_coul_ent,
-                             cpt_elt_coul_ent);
+                             NULL,
+                             NULL,
+                             NULL);
 }
 
 /*============================================================================
@@ -1649,6 +1947,11 @@ ecs_pre_gmsh__lit_maillage(const char  *nom_fic_maillage)
   bool          bool_elements = false;
   bool          bool_noeuds = false;
 
+  int           n_phys_names = 0;
+  int           n_phys_names_max = 0;
+  int          *phys_name_id = NULL;
+  char        **phys_name = NULL;
+
   /* Création d'un maillage initialement vide (valeur de retour) */
 
   ecs_maillage_t  *maillage = ecs_maillage__cree_nodal();
@@ -1671,6 +1974,8 @@ ecs_pre_gmsh__lit_maillage(const char  *nom_fic_maillage)
   num_ligne = 0;
 
   som_val_label = NULL;
+
+  int fam_shift[ECS_N_ENTMAIL] = {0, 0};
 
   /* Par défaut, la version est 1.0 en ASCII */
 
@@ -1701,6 +2006,33 @@ ecs_pre_gmsh__lit_maillage(const char  *nom_fic_maillage)
                                         &version_fmt_gmsh,
                                         &type_fmt_gmsh,
                                         &num_ligne);
+
+    else if (strncmp(chaine,
+                     "$PhysicalNames",
+                     strlen("$PhysicalNames")) == 0)
+      _read_physical_names(fic_maillage,
+                           &num_ligne,
+                           &n_phys_names,
+                           &n_phys_names_max,
+                           &phys_name_id,
+                           &phys_name);
+
+    else if (strncmp(chaine,
+                     "$Entities",
+                     strlen("$Entities")) == 0) {
+
+      ecs_famille_t  *vect_famille[ECS_N_ENTMAIL];
+      _read_entities(n_phys_names_max,
+                     phys_name_id,
+                     (const char**)phys_name,
+                     fic_maillage,
+                     &num_ligne,
+                     type_fmt_gmsh,
+                     fam_shift,
+                     vect_famille);
+
+      ecs_maillage__definit_famille(maillage, vect_famille);
+    }
 
     /* Si la chaine lue marque le début de la section des noeuds */
 
@@ -1747,6 +2079,7 @@ ecs_pre_gmsh__lit_maillage(const char  *nom_fic_maillage)
                                           type_fmt_gmsh,
                                           &num_ligne,
                                           maillage->n_vertices,
+                                          fam_shift,
                                           &som_val_label);
       else
         ecs_loc_pre_gmsh__lit_elements(maillage,
@@ -1791,6 +2124,18 @@ ecs_pre_gmsh__lit_maillage(const char  *nom_fic_maillage)
   /*----------------------------------*/
 
   ecs_file_free(fic_maillage);
+
+  /* Transformation des familles en attributs "groupe" */
+  /*---------------------------------------------------*/
+
+  if (version_fmt_gmsh == 4)
+    ecs_maillage__cree_attributs(maillage);
+
+  for (int i = 0; i < n_phys_names; i++)
+    ECS_FREE(phys_name[i]);
+
+  ECS_FREE(phys_name);
+  ECS_FREE(phys_name_id);
 
   /* Retour */
   /*========*/